Fx double server file
