+++ package.json
@@ -24,7 +24,7 @@
   },
   "scripts": {
     "start": "node -r module-alias/register dist/server.js",
-    "dev": "ts-node-dev -r tsconfig-paths/register --respawn --transpile-only server.ts",
+    "dev": "ts-node-dev -r tsconfig-paths/register --respawn --transpile-only src/server.ts",
     "build": "tsc && tsc-alias",
     "test": "jest --detectOpenHandles",
     "test:watch": "jest --watch --detectOpenHandles",
