+++ src/modules/order/domain/order.entity.ts
@@ -1,8 +1,7 @@
-import mongoose from 'mongoose';
-import { ICustomer } from '@modules/customer/domain/customer.entity'; // Import Customer
-import { IVoucher } from '@modules/voucher/domain/voucher.entity';
-import { PaymentDetails, IPaymentDetails } from './order.interface'; // Corrected import for PaymentDetails
 import { Types } from 'mongoose';
+import { ICustomer } from '@modules/customer/domain/customer.entity';
+import { IVoucher } from '@modules/voucher/domain/voucher.entity';
+import { IPaymentDetails } from './order.interface';
 
 export enum OrderStatus {
   PENDING = 'pending',
@@ -13,22 +12,24 @@ export enum OrderStatus {
 }
 
 export interface IOrder {
-  _id?: mongoose.Types.ObjectId;
-  // userId?: mongoose.Types.ObjectId | null; // Removed userId
-  customerId: mongoose.Types.ObjectId | ICustomer; // Added customerId (required)
-  paymentDetails: IPaymentDetails; // Use IPaymentDetails type
-  voucher: mongoose.Types.ObjectId | IVoucher; // Link to the generated voucher
+  _id?: Types.ObjectId;
+  customerId: Types.ObjectId | ICustomer;
+  paymentDetails: IPaymentDetails;
+  voucher: Types.ObjectId | IVoucher;
   status: OrderStatus;
+  emailsSent: boolean;
+  pdfGenerated: boolean;
+  pdfUrl?: string;
   createdAt?: Date;
   updatedAt?: Date;
 }
 
 export class Order implements IOrder {
   _id?: Types.ObjectId;
-  // userId?: mongoose.Types.ObjectId | null;
-  customerId: string | Types.ObjectId;
-  voucher: any; // TODO: Add proper voucher type
+  customerId: Types.ObjectId | ICustomer;
+  voucher: Types.ObjectId | IVoucher;
   paymentDetails: IPaymentDetails;
+  status: OrderStatus;
   emailsSent: boolean;
   pdfGenerated: boolean;
   pdfUrl?: string;
@@ -37,10 +38,10 @@ export class Order implements IOrder {
 
   constructor(data: IOrder) {
     this._id = data._id;
-    // this.userId = data.userId;
     this.customerId = data.customerId;
     this.voucher = data.voucher;
     this.paymentDetails = data.paymentDetails;
+    this.status = data.status;
     this.emailsSent = data.emailsSent;
     this.pdfGenerated = data.pdfGenerated;
     this.pdfUrl = data.pdfUrl;
