+++ src/server.ts
@@ -4,7 +4,6 @@ import './bootstrap';
 import express, { Request, Response, NextFunction, ErrorRequestHandler } from 'express';
 import passport from 'passport';
 import dotenv from 'dotenv';
-import mongoose from 'mongoose';
 import path from 'path';
 import fs from 'fs';
 import cors from 'cors';
@@ -16,31 +15,40 @@ import { connectDB } from '@shared/infrastructure/database/connection';
 import logger from '@shared/infrastructure/logging/logger';
 import { initCloudinary } from '@shared/infrastructure/services/cloudinary.config';
 import { AppError, ErrorTypes } from '@shared/types/appError';
-import { errorHandler as globalErrorHandler } from '@shared/infrastructure/errors/errorHandler';
+import { errorHandler as globalErrorHandler, notFoundHandler } from '@shared/infrastructure/errors/errorHandler';
 import { setupSwagger } from '@shared/infrastructure/swagger/swagger';
 import { applySecurityMiddleware } from '@shared/infrastructure/middleware/security';
-import { setupRoutes } from './modules/routes';
+import { setupRoutes } from '@modules/routes';
+import { authenticate } from '@shared/infrastructure/middleware/auth';
 
-// Import routes
+// Import routes and controllers
 import userRoutes from '@modules/user/interface/user.routes';
 import passwordResetRoutes from '@modules/user/interface/passwordReset.routes';
 import { storeRoutes } from '@modules/store/interface/store.routes';
 import { productRoutes } from '@modules/product/interface/product.routes';
 import voucherRoutes from '@modules/voucher/interface/voucher.routes';
 import orderRoutes from '@modules/order/interface/order.routes';
-const customerRoutes = require('@modules/customer/interface/customer.routes');
+
 import { UserController } from '@modules/user/interface/user.controller';
 import { UserService } from '@modules/user/application/user.service';
 import { MongoUserRepository } from '@modules/user/infrastructure/user.repository';
+
+import { CustomerController } from '@modules/customer/interface/customer.controller';
+import { CustomerService } from '@modules/customer/application/customer.service';
+import { CustomerRepository } from '@modules/customer/domain/customer.repository';
 
 dotenv.config();
 
 const app = express();
 
-// Initialize user controller for login route
+// Initialize controllers
 const userRepository = new MongoUserRepository();
 const userService = new UserService(userRepository);
 const userController = new UserController(userService);
+
+const customerRepository = new CustomerRepository();
+const customerService = new CustomerService(customerRepository);
+const customerController = new CustomerController(customerService);
 
 // Apply middleware
 app.use(cors());
@@ -49,6 +57,24 @@ app.use(compression());
 app.use(morgan('dev'));
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
+app.use(passport.initialize());
+
+// Create uploads directory for voucher PDFs
+const uploadsDir = path.join(__dirname, '../uploads');
+const vouchersDir = path.join(uploadsDir, 'vouchers');
+
+if (!fs.existsSync(uploadsDir)) {
+  logger.info(`Creating uploads directory: ${uploadsDir}`);
+  fs.mkdirSync(uploadsDir, { recursive: true });
+}
+
+if (!fs.existsSync(vouchersDir)) {
+  logger.info(`Creating vouchers directory: ${vouchersDir}`);
+  fs.mkdirSync(vouchersDir, { recursive: true });
+}
+
+// Serve static files from uploads directory
+app.use('/uploads', express.static(path.join(__dirname, '../uploads')));
 
 // Initialize services
 connectDB()
@@ -68,7 +94,6 @@ setupRoutes(app);
 // Public routes
 app.post('/api/v1/login', (req: Request, res: Response, next: NextFunction) => userController.login(req, res, next));
 app.use('/api/v1/auth', passwordResetRoutes);
-app.use('/api/v1/customers', customerRoutes);
 
 // Health check endpoint
 app.get('/health', (req: Request, res: Response) => {
@@ -82,33 +107,48 @@ app.use('/api/v1/products', productRoutes);
 app.use('/api/v1/vouchers', voucherRoutes);
 app.use('/api/v1/orders', orderRoutes);
 
+// Customer routes
+app.get('/api/v1/customers-test', (req: Request, res: Response) => {
+  res.status(200).json({ message: 'Customer test route is working!' });
+});
+
+app.get('/api/v1/customers', authenticate, (req: Request, res: Response, next: NextFunction) => {
+  customerController.getCustomers(req, res, next);
+});
+
+app.post('/api/v1/customers', (req: Request, res: Response, next: NextFunction) => {
+  customerController.createCustomer(req, res, next);
+});
+
+app.post('/api/v1/customers/get-or-create', (req: Request, res: Response, next: NextFunction) => {
+  customerController.getOrCreateCustomer(req, res, next);
+});
+
+app.get('/api/v1/customers/:id', authenticate, (req: Request, res: Response, next: NextFunction) => {
+  customerController.getCustomerById(req, res, next);
+});
+
+app.put('/api/v1/customers/:id', authenticate, (req: Request, res: Response, next: NextFunction) => {
+  customerController.updateCustomer(req, res, next);
+});
+
+app.delete('/api/v1/customers/:id', authenticate, (req: Request, res: Response, next: NextFunction) => {
+  customerController.deleteCustomer(req, res, next);
+});
+
 // Setup Swagger documentation
 setupSwagger(app);
 
-// Error handling middleware
-const errorHandler: ErrorRequestHandler = (
-  err: AppError,
-  req: Request,
-  res: Response,
-  _next: NextFunction
-): void => {
-  logger.error('Error:', err);
+// Add 404 handler
+app.use(notFoundHandler);
 
-  if (err.isOperational) {
-    res.status(err.status || 500).json({
-      status: err.status || 'error',
-      message: err.message,
-    });
-    return;
-  }
+// Global error handler
+app.use(globalErrorHandler);
 
-  // Programming or other unknown error
-  res.status(500).json({
-    status: 'error',
-    message: 'Something went wrong!',
-  });
-};
-
-app.use(errorHandler);
+// Start server
+const PORT = process.env.PORT || 3000;
+app.listen(PORT, () => {
+  logger.info(`Server is running on port ${PORT}`);
+});
 
 export default app;
